Rake task to check service flow object creation
The service flow objects should be created automatically when a user
visits the flow page for their service.

This tasks checks that each service has had a flow object created and
that it is correct.

Since there is no branching at this time each flow object should link to
the next page at the correct index in the pages array except for the
last page which should just be an empty string