Create service flow objects rake task
Older services made use of the pages array their flow as since there was
no conditional logic the editor and runner would just traverse the array
in order.

The runner is backwards compatible and will make use of either flow
objects or an array of page objects.

This adds a task to create the flow objects from the pages
array for services that do not already use the flow objects.

There is already a before_action in the editor to do this for any user
that visits the flow page for their form. This task is a backup for
that.