Add GET endpoint to get items associated with a component
To minimise the amount of data that is returned to the Editor, we
decided to create a new endpoint that returns items associated with
one component.

This allows us to check whether there are items already associated with
the one component, which is something we need to cater for in the
Editor.