Cater for detached autocomplete components
Prior to this change we would return all the items associated with
an autocomplete component, this included items that were associated
with a detached component. We do not want to return/inject any
unnecessary items into the container.

This change ensures we only get items that are associated with
components in the main flow of a form.
Since we now rely on the grid and flow objects, we need to ensure our
test metadata accurately reflects the metadata in a live form, hence
we have decided to use a fixture.

The fixture contains 4 autocomplete components, one of which is detached.

Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>
Co-authored-by: Hellema Ibrahim <hellema.ibrahim@digital.justice.gov.uk>